Allow the selection of different graphics drivers
On my new Tuxedo laptop I had the problem that it has an intel graphics
card, but the I cannot use the intel graphics driver (it caused
problems). Therefore the possibility to select the driver in the
coresponding xorg config file.

This involved moving the 00-keyboard.conf file as well (into the
template directory). It is not really a template, but otherwise I would
have to differentiate between templates and "normal" files in the x11
task. And who know, maybe it will become a template in the future.